requirements.txt
fastapi
aiohttp
uvicorn
httpx
slack_bolt
slack-click
tenacity